+++ src/main/java/net/silentchaos512/gear/network/Network.java
@@ -54,6 +54,11 @@ public final class Network {
                 .decoder(SyncGearCraftingItemsPacket::fromBytes)
                 .consumer(SyncGearCraftingItemsPacket::handle)
                 .add();
+        channel.messageBuilder(ShowPartsScreenPacket.class, 5)
+                .encoder((packet, buffer) -> {})
+                .decoder(buffer -> new ShowPartsScreenPacket())
+                .consumer(ShowPartsScreenPacket::handle)
+                .add();
     }
 
     private Network() {}