+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.7] - 2020-04-12
 ### Added
 - Command to show parts GUI (`/sgear_parts show_gui`). Does not require OP perms. [#107]
 - Crimson iron and steel blocks [#88]